skif-elo/requirements.txt
nextcord
mysql.connector
os